# openAPI.yaml
openapi: 3.0.0
info:
  title: Certificate Generation API
  description: API for generating and managing certificates
  version: 1.0.0
servers:
  - url: http://localhost:3000
paths:
  /certificate:
    get:
      summary: Get certificate PDF
      description: Retrieves a certificate PDF by request ID
      parameters:
        - in: query
          name: requestId
          description: ID of the certificate request
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Certificate request data not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /liveness:
    get:
      summary: Health check endpoint
      description: Endpoint for checking the liveness of the server
      responses:
        '200':
          description: OK
        '500':
          description: Internal server error
  /readiness:
    get:
      summary: Readiness check endpoint
      description: Endpoint for checking the readiness of the server
      responses:
        '200':
          description: OK
        '404':
          description: Certificate request data not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string